Continue to refactor Drop extra lias
