+++ app/hooks/useGetContent.tsx
@@ -1,7 +1,7 @@
 import { useEffect, useState } from "react";
 
-import { HnItem, useDataStore } from "@/stores/useDataStore";
 import { getColorsForStory } from "~/lib/getColorsForStory";
+import { HnItem, useDataStore } from "~/stores/useDataStore";
 
 export function useGetContent(id: number, initialSsrData: HnItem | undefined) {
   const [storyData, setStoryData] = useState<HnItem | undefined>(
